Only remove the local work dir if it is present

+++ ._wb/tool/download_genomes/run.sh
@@ -17,9 +17,13 @@ nextflow \
     -resume \
     -profile "${PROFILE}"
 
-# Delete the temporary files created during execution
-echo Removing temporary files
-rm -r work
+# If temporary files were not placed in a separate location
+if [ -d work ]; then
+    # Delete the temporary files created during execution
+    echo Removing temporary files
+    rm -r work
+fi
+
 
 echo
 date
